Correctly substitute generic value classes in signatures
^KT-52706 Fixed

We need to carefully track whether we are producing a type for a signature, and in that case we should not use the unsubstituted version of the underlying type of the value class, but the substituted one. Except that we need to take special care of the case in which we always use Object as the representation and we should not change that.

Apply suggestions