IR deserializer: Use UNDEFINED_OFFSET for offsets of deserialized IR annotations
The current IR proto format does not allow serializing offsets for
IR annotations. Earlier, during the deserialization phase, zero (0) was
used as a value for offsets. That's not correct, because 0 is still
a valid offset inside of IR file.

It's better to use here UNDEFINED_OFFSET to indicate that the offset
is not known.