[FIR] reproduce incorrect session cache
`jkjSyntheticProperty.kt` reproduces the issue.
The issue is caused by an incorrectly configured key
for `ScopeSession#getOrBuild`. In particular,
`buildJavaEnhancementScope` accepts `memberRequiredPhase`,
but the key doesn't depend on it. As a result,
someone might request a scope without `memberRequiredPhase` and
effectively classes would be resolved up to the super types phase
(or any other phase). After that another client requested
scope with `memberRequiredPhase` (the status phase is our case),
but the previous cached result is returned, which is wrong.
Side note: it is crucial to have the `MultiPlatformProjects` flag
in the test as it triggers logic inside `FirDefaultParametersResolver`
for `FirActualizingScope`, which calls `scopeForClass` without
the required phase

^KT-75323
^KT-75366