[FIR] Skip implicit when subjects in `isAppropriateTarget()`
Otherwise, `K2IntroduceFunctionTestGenerated.testEvalExprInWhenSubject`
fails in `intellij.git`: it checks an action extracting a when subject
expression to a separate function, which is now performed as if we've
selected a variable declaration instead of just the subject expression.

It's important to fix it as it used to work previously (in K2 IDE and K1).
Note that IDE will still generate incorrect extraction if asked to
create it for an explicit when subject, but this is not a big problem,
as K1 IDE also works this way.

For `when (a + b) {`, The code produced is:

```
fun foo(a: Int): Int {
    val b: Int = 1
    return when (__dummyTestFun__(a, b)) {
        0 -> b
        1 -> -b
        else -> a - b
    }
}

private fun __dummyTestFun__(a: Int, b: Int) {
    a + b
}
```

(But in the case of explicit when subjects, there's an explicit
`val subject = ...` in `__dummyTestFun__()`.